Former-commit-id: 2fffcde8cf6e0abc379b81457bac5818e6546e79

+++ instrumented_test.py
@@ -200,8 +200,11 @@ class Run:
         if not os.path.exists(dir_name_to_save_results_xml):
             os.makedirs(dir_name_to_save_results_xml)
 
-        shutil.move("build/reports/androidTests", dir_name_to_save_results_html)
-        shutil.move("id/build/outputs/androidTest-results/connected", dir_name_to_save_results_xml)
+        if os.path.exists("build/reports/androidTests"):
+            shutil.move("build/reports/androidTests", dir_name_to_save_results_html)
+
+        if os.path.exists("id/build/outputs/androidTest-results/connected"):
+            shutil.move("id/build/outputs/androidTest-results/connected", dir_name_to_save_results_xml)
 
 
 def main():
@@ -209,7 +212,6 @@ def main():
 
     run = Run('instrumented_test')
     run.update_log_format(LogState.default())
-    run.log("Hello world!")
 
     run.clean_cerberus_build()
     run.assemble_cerberus_apk()
